rearange files to get structure for pytest, rename Current Tests to pytest pattern, add init-file to include solve in pyFiles
