+++ .gitignore
@@ -21,3 +21,4 @@ build
 Non-Lcapy-Files/Solutions
 .idea
 dist
+NonLcapyFiles/Solutions
