+++ NonLcapyFiles/StandardCircuits/Circuit_capacitors.txt
