+++ NonLcapyFiles/StandardCircuits/Circuit_inductors.txt
