+++ NonLcapyFiles/StandardCircuits/Circuit_mixed.txt
