+++ NonLcapyFiles/StandardCircuits/Circuit_mixed_2pi30.txt
