+++ NonLcapyFiles/StandardCircuits/Circuit_mixed_30.txt
