+++ NonLcapyFiles/StandardCircuits/Circuit_mixed_omega0.txt
