+++ NonLcapyFiles/StandardCircuits/Circuit_resistors.txt
